4.21 completed

+++ controllers/blogs.js
@@ -46,8 +46,24 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, config.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'token invalid'
+    });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  if ( blog.user.toString() === user.id){
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(400).json({
+      error: 'User not authorized to delete this blog post.'
+    })
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
